% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everything.R
\docType{data}
\name{apiKey}
\alias{apiKey}
\title{Get all resources that the API of newsapi.org provides}
\format{An object of class \code{character} of length 1.}
\usage{
apiKey
}
\arguments{
\item{api_key}{Character string with the API key you get from newsapi.org. Passing it is compulsory. Function gets it per default from your global environment. See "set-api-key" for more info.}

\item{content}{Character string that contains the content of your search in the API data base}

\item{nrtexts}{The number (numeric!) of articles per page that are returned. Maximum is 100, default is 20.}

\item{sources}{Character string with IDs (comma separated) of the news outlets you want to focus on (e.g., "usa-today, spiegel-online").}

\item{domains}{Character string (comma separated) with domains that you want to restrict your search to (e.g. "bbc.com, nytimes.com").}

\item{exclude_domains}{Similar usage as 'domains'. Will exclude these domains from your search.}

\item{from}{Marks the date of the start of your search. Must be in ISO 8601 format (e.g., '2018-09-08' or '2018-09-08T12:51:42'). Default is the oldest.}

\item{to}{Marks the end date of your search. Works similarly as 'from'. Default is the latest article.}

\item{language}{Specifies the language of the articles of your search. Must be in ISO shortcut format (e.g., "de", "en"). See list of all languages on https://newsapi.org/docs/endpoints/everything. Default is all languages.}

\item{sort_by}{Character string that specifies the sorting of your article results. Has three options: "publishedAt", "relevancy", "popularity". Default is "publishedAt".}

\item{page}{Specifies the page number of your results. Must be numeric. Default is first page.}
}
\value{
A list containing a data frame with results for your search and another list with metadata on your search.
}
\description{
Get all resources that the API of newsapi.org provides
}
\examples{
get_everything(api_key = key, content = "stuttgart", language = "de")
get_everything(api_key = key, content = "mannheim", from = "2018-01-07T12:00:00")

}
\keyword{datasets}
