% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_newsanchor_url}
\alias{build_newsanchor_url}
\title{Builds query URL for newsapi.org.}
\usage{
build_newsanchor_url(url, query_args)
}
\arguments{
\item{url}{NEWS API endpoint.}

\item{query_params}{named list of parameters that are needed to query the endpoint. 
Check the News API documentation to see which endpoint requires which parameters.}
}
\value{
httr URL.
}
\description{
\code{build_newsanchor_url} adds a list of query arguments to a given 
News API endpoint.
}
