% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_to_comma_separated}
\alias{collapse_to_comma_separated}
\title{Concatenate character vector to comma-separated string.}
\usage{
collapse_to_comma_separated(v)
}
\arguments{
\item{v}{character vector.}
}
\value{
string with elements of v separated by comma.
}
\description{
\code{collapse_to_comma_separated} is a helper function that concatenates a character vector 
to a comma-separated string. If the input vector has only one element, the element will be returned unchanged.
}
