% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_searchterms.R
\name{create_data_searchterms}
\alias{create_data_searchterms}
\title{create_data_searchterms}
\usage{
create_data_searchterms()
}
\description{
This script creates three different dataframes with possible search terms 
for 'country', 'category', and news 'sources'. This function should be 
ignored from building the package, but rather allows to update the dataframes
of available searchterms.
}
\examples{
create_list_searchterms()
}
