% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_searchterms.R
\name{create_list_searchterms}
\alias{create_list_searchterms}
\title{create_list_searchterms}
\usage{
create_list_searchterms(api_key = NULL)
}
\arguments{
\item{api_key}{API key required to access 'sources'}
}
\description{
This function creates a list with possible search terms for 'country',
'category', and news 'sources'. This is not intended to be distributed in
the package, but rather allows to update the dataset of searchterms.
}
\examples{
create_list_searchterms(api_key)
create_list_searchterms()
}
