% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_newsanchor_articles}
\alias{extract_newsanchor_articles}
\title{Extracts data frame with News API articles from response object.}
\usage{
extract_newsanchor_articles(metadata, content_parsed)
}
\arguments{
\item{metadata}{data frame containing meta data related to the request, see extract_newsanchor_metadata.}

\item{content_parsed}{parsed content of a response to News API query}
}
\value{
data frame containing articles.
}
\description{
\code{extract_newsanchor_articles} extracts a data frame containing the News API articles that
matched the request to News API everything or headlines endpoint.
}
