% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_newsanchor_metadata}
\alias{extract_newsanchor_metadata}
\title{Extracts metadata.}
\usage{
extract_newsanchor_metadata(response, content_parsed, page = NULL,
  page_size = NULL)
}
\arguments{
\item{response}{httr response object}

\item{content_parsed}{parsed content of a response to News API query}

\item{page}{Specifies the page number of your results that was returned. Defaults to NULL.}

\item{page_size}{The number of articles per page that were returned. Defaults to NULL.}
}
\value{
data frame containing meta data related to the query.
}
\description{
\code{extract_newsanchor_metadata} extracts meta data from the response object and the 
parsed content.
}
