% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_newsanchor_sources}
\alias{extract_newsanchor_sources}
\title{Extracts data frame with News API sources from response object.}
\usage{
extract_newsanchor_sources(metadata, content_parsed)
}
\arguments{
\item{metadata}{data frame containing meta data related to the request, see extract_newsanchor_metadata.}

\item{content_parsed}{parsed content of a response to News API query}
}
\value{
data frame containing sources.
}
\description{
\code{extract_newsanchor_sources} extracts a data frame containing the News API sources that
matched the request to News API sources endpoint.
}
