% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headlines.R
\name{get_headlines}
\alias{get_headlines}
\title{Returns selected headlines from newsapi.org}
\usage{
get_headlines(query = NULL, category = NULL, country = NULL,
  sources = NULL, page = 1, page_size = 100,
  api_key = Sys.getenv("NEWS_API_KEY"))
}
\arguments{
\item{query}{Keyword you want headlines for}

\item{category}{Category you want headlines from}

\item{country}{Country you want headlines for}

\item{sources}{Sources you want headlines from}

\item{page}{Use this to page through the results, if the total results 
found is greater than the \code{page_size}.}

\item{page_size}{Number of hits per request (maximum = 100)}

\item{api_key}{Your API token}
}
\value{
List with two dataframes:\cr
        1) Data frame with \code{results_df}\cr
        2) Data frame with \code{meta_data}
}
\description{
\code{get_headlines} returns live top and breaking headlines for a country, 
specific category in a country, single source, or multiple sources. You can 
also search with keywords. Articles are sorted by the earliest date 
published first. To automatically download all results, use 
\code{get_headlines_all}\cr\cr
Please check that the api_key is available. You can provide an explicit
definition of the api_key or use \code{set_api_key} \cr\cr
For valid searchterms see \code{data(searchterms)}
}
\examples{
\dontrun{
df <- get_headlines(sources = "bbc-news")
}
}
