% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_newsanchor_get_request}
\alias{make_newsanchor_get_request}
\title{Makes a GET request to News API.}
\usage{
make_newsanchor_get_request(url, api_key)
}
\arguments{
\item{url}{News API url with query parameters and scheme specified. See build_newsanchor_url.}

\item{api_key}{News API key.}
}
\value{
httr response object.
}
\description{
\code{make_newsanchor_get_request} makes a GET request to News API.
}
