% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_newsanchor_content}
\alias{parse_newsanchor_content}
\title{Parses content returned by query to the News API.}
\usage{
parse_newsanchor_content(response)
}
\arguments{
\item{response}{httr response object}
}
\value{
R list.
}
\description{
\code{parse_newsanchor_content} parses the content sent back by
the News API to an R list.
}
