% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headlines.R
\name{\%>\%}
\alias{\%>\%}
\alias{get_headlines}
\title{get_headlines}
\usage{
get_headlines(api_key, sources = NULL, country = NULL, category = NULL,
  keyword = NULL, pageSize = 100, page = 1)
}
\arguments{
\item{api_key}{Character string API token.}

\item{country}{Name of country}

\item{category}{Name of category you want headlines for}

\item{keyword}{Name potential keyword}

\item{pageSize}{Number of hits per request}

\item{page}{Number of page per request}

\item{source}{Name of news source.}
}
\value{
Data frame
}
\description{
Returns headlines from news sources.
}
\examples{
\dontrun{
df <- get_headlines(api_key, "bbc-news")
}
}
