% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_response.R
\docType{data}
\name{sample_response}
\alias{sample_response}
\title{Sample Response Object}
\format{An object of class \code{list} of length 2.}
\usage{
sample_response
}
\value{
List with two dataframes:\cr
        1) Data frame with \code{results_df}\cr
        2) Data frame with \code{meta_data}
}
\description{
A sample response object generated using `get_everything`.
}
\details{
This response object was mainly created for demonstrating purposes.
The data set is used in the "Scrape New York Times Online Articles" vignette.
The object was created using the following query.
}
\examples{
\dontrun{
response <- get_everything(query   = "Trump",
                           sources = "the-new-york-times",
                           from    = "2018-12-03",
                           to      = "2018-12-09") 
}
}
\keyword{datasets}
