% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_invalid_category}
\alias{stop_if_invalid_category}
\title{Checks validity of a category.}
\usage{
stop_if_invalid_category(category)
}
\arguments{
\item{category}{category to check as a string.}
}
\description{
\code{stop_if_invalid_category} checks whether a given category is valid for News API and 
stops with an error if this is not the case.
}
