% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_invalid_country}
\alias{stop_if_invalid_country}
\title{Checks validity of a country}
\usage{
stop_if_invalid_country(country)
}
\arguments{
\item{country}{country to check as a string.}
}
\description{
\code{stop_if_invalid_country} checks whether a given country is valid for News API and 
stops with an error if this is not the case.
}
