% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_invalid_language}
\alias{stop_if_invalid_language}
\title{Checks validity of a language}
\usage{
stop_if_invalid_language(language)
}
\arguments{
\item{language}{language to check as a string.}
}
\description{
\code{stop_if_invalid_language} checks whether a given language is valid for News API and 
stops with an error if this is not the case.
}
