% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_invalid_source}
\alias{stop_if_invalid_source}
\title{Checks validity of a source}
\usage{
stop_if_invalid_source(source)
}
\arguments{
\item{source}{source to check as a string.}
}
\description{
\code{stop_if_invalid_source} checks whether a given source is valid for News API and 
stops with an error if this is not the case.
}
