% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_category.R
\docType{data}
\name{terms_category}
\alias{terms_category}
\title{Terms Category}
\format{
An object of class \code{data.frame} with 7 rows and 1 columns.
}
\usage{
terms_category
}
\description{
The dataframe `provides possible categories (e.g., sports) you want to get 
headlines for. This dataframe is relevant in conjunction with 
\code{get_headlines}.
}
\keyword{datasets}
