% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_language.R
\docType{data}
\name{terms_language}
\alias{terms_language}
\title{Terms Language}
\format{An object of class \code{data.frame} with 14 rows and 1 columns.}
\usage{
terms_language
}
\description{
Terms Language
}
\keyword{datasets}
