% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_sources.R
\docType{data}
\name{terms_sources}
\alias{terms_sources}
\title{Terms Sources}
\format{
An object of class \code{data.frame} with 138 rows and 1 columns.
}
\usage{
terms_sources
}
\description{
This dataframe provides possible news sources or blogs you want
to get news from. This dataframe is relevant in conjunction with 
\code{get_everything}.
}
\keyword{datasets}
