% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sources.R
\name{get_sources}
\alias{get_sources}
\title{Get (currently) available sources from newsapi.org.}
\usage{
get_sources(apiKey = NULL, category = NULL, language = NULL,
  country = NULL)
}
\arguments{
\item{apiKey}{your API key - it is required to make the API call.}

\item{category}{any category you might want to explore. Must be character or string}

\item{language}{the language of the newsoutlet (e.g. "de" or "en").}

\item{country}{the newsoutlet's country (e.g. "us").}
}
\value{
A dataframe based on your specifications
}
\description{
Get (currently) available sources from newsapi.org.
}
\examples{
get_sources(api)
get_sources(api, category="technology")
get_sources(api, language="en")
}
