% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_category.R
\docType{data}
\name{terms_category}
\alias{terms_category}
\title{Terms Category}
\format{An object of class \code{data.frame} with 7 rows and 1 columns.}
\usage{
terms_category
}
\description{
Terms Category
}
\keyword{datasets}
