% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_country.R
\docType{data}
\name{terms_country}
\alias{terms_country}
\title{Terms Country}
\format{An object of class \code{data.frame} with 54 rows and 1 columns.}
\usage{
terms_country
}
\description{
Terms Country
}
\keyword{datasets}
